Prevent user for deleting his own account
An administrator cannot disable or delete his own account and he cannot suppress admin rights from his own account.
These operations have to be done by another administrator.

+++ application/modules/admin/controllers/Admin.php
@@ -143,10 +143,9 @@ class Admin extends MY_Controller
             }
             redirect('admin/user_index');
         } else {
-            $old_values = [
-                'user_name' => $this->input->post('user_name'),
-                'user_usertype' => $this->input->post('user_usertype')
-            ];
+            $old_values['user_name'] = $this->input->post('user_name');
+            if($_SESSION['user_id'] != $user_id)
+                $old_values['user_usertype'] = $this->input->post('user_usertype');
             $this->user_add($user_id, $old_values);
         }
     }
@@ -164,7 +163,9 @@ class Admin extends MY_Controller
     public function user_delete($user_id, $action = 0)
     {
         $user = $this->user_model->with_deleted()->get($user_id);
-        if (is_null($user)) redirect('admin/user_index');
+        if (is_null($user)) {
+            redirect('admin/user_index');
+        }
 
         switch($action) {
             case 0: // Display confirmation
@@ -175,10 +176,14 @@ class Admin extends MY_Controller
                 $this->display_view('admin/user/delete', $output);
                 break;
             case 1: // Deactivate (soft delete) user
-                $this->user_model->delete($user_id, FALSE);
+                if ($_SESSION['user_id'] != $user->id) {
+                    $this->user_model->delete($user_id, FALSE);
+                }
                 redirect('admin/user_index');
             case 2: // Delete user
-                $this->user_model->delete($user_id, TRUE);
+                if ($_SESSION['user_id'] != $user->id) {
+                    $this->user_model->delete($user_id, TRUE);
+                }
                 redirect('admin/user_index');
             default: // Do nothing
                 redirect('admin/user_index');
